+++ src/Server/staradmin/Commands/CommandFactory.cs
@@ -57,11 +57,13 @@ namespace staradmin.Commands {
                 string serverHost;
                 int serverPort;
                 string ignored;
-                SharedCLI.ResolveDatabase(args, out database);
+
+                var explicitDatabase = SharedCLI.ResolveDatabase(args, out database);
                 SharedCLI.ResolveAdminServer(args, out serverHost, out serverPort, out ignored);
  
                 var context = new Context(args);
                 context.Database = database;
+                context.IsExcplicitDatabase = explicitDatabase;
                 context.ServerHost = serverHost;
                 context.ServerPort = serverPort;