+++ src/Starcounter.Installer/Starcounter.InstallerWPF/Configuration.cs
@@ -426,8 +426,18 @@ namespace Starcounter.InstallerWPF {
                         }
                     }
                     else {
+
+                        CheckforEmptyFolder:
+                        int checks = 6;
                         // Checking if there are any files in the target installation directory.
                         if (Utilities.DirectoryIsNotEmpty(new DirectoryInfo(installationPath))) {
+
+                            Thread.Sleep(1000);
+                            checks--;
+                            if (checks > 0) {
+                                goto CheckforEmptyFolder;
+                            }
+
                             // Setting normal attributes for all files and folders in installation directory.
                             Utilities.SetNormalDirectoryAttributes(new DirectoryInfo(installationPath));