+++ src/Starcounter.Installer/Starcounter.InstallerWPF/InitializationWindow.xaml.cs
@@ -249,8 +249,7 @@ namespace Starcounter.InstallerWPF {
 
                 WpfMessageBoxResult userChoice = WpfMessageBoxResult.None;
 
-                String uninstallQuestion = string.Format("Do you want to {0} from version {1} to {2} ?", (ScVersionDate > installedVersionDate) ? "upgrade" : "downgrade", installedVersion, ScVersion),
-                    headingMessage = "Starcounter Installation";
+                String upgradeQuestion = string.Format("Would you like to {0} Starcounter v{1} to v{2}?", (ScVersionDate > installedVersionDate) ? "upgrade" : "downgrade", installedVersion, ScVersion), headingMessage = "Starcounter Installation";
 
                 // Checking for the existing databases compatibility.
                 List<String> dbListToUnload = new List<String>();
@@ -269,7 +268,7 @@ namespace Starcounter.InstallerWPF {
 
                         String dbListToUnloadText = String.Join(Environment.NewLine, dbListToUnload);
 
-                        uninstallQuestion += Environment.NewLine + Environment.NewLine +
+                        upgradeQuestion += Environment.NewLine + Environment.NewLine +
                             "Existing database image files are incompatible with this installation (database(s): " + dbListToUnloadText + "). " +
                             "Please follow the instructions at: " + Environment.NewLine +
                             "https://github.com/Starcounter/Starcounter/wiki/Reloading-database-between-Starcounter-versions " + Environment.NewLine +
@@ -280,7 +279,7 @@ namespace Starcounter.InstallerWPF {
                 else {
 
                     // Some error occurred during the check.
-                    uninstallQuestion +=
+                    upgradeQuestion +=
                         "Error occurred during verification of existing database image files versions." + Environment.NewLine +
                         "Please follow the instructions at: " + Environment.NewLine +
                         "https://github.com/Starcounter/Starcounter/wiki/Reloading-database-between-Starcounter-versions " + Environment.NewLine +
@@ -294,11 +293,16 @@ namespace Starcounter.InstallerWPF {
 
                 // Asking for user choice about uninstalling.
                 userChoice = WpfMessageBox.Show(
-                    uninstallQuestion,
+                    upgradeQuestion,
                     headingMessage,
                     WpfMessageBoxButton.YesNo, WpfMessageBoxImage.Question);
 
                 if (userChoice == WpfMessageBoxResult.Yes) {
+
+                    this.isUpgrade = true;
+                    this.unattended = true;
+                    this.setupOptions = SetupOptions.Install;
+                    this.finishedMessage = string.Format("{0} successful.", (ScVersionDate > installedVersionDate) ? "Upgrade" : "Downgrade");
 
                     // Asking to launch current installed version uninstaller.
                     String installDir = GetInstalledDirFromEnv();
@@ -407,7 +411,6 @@ namespace Starcounter.InstallerWPF {
             Element_Storyboard.Stop(this.PART_Canvas);
         }
 
-
         void InitializationWindow_Loaded(object sender, RoutedEventArgs e) {
 
             // TODO: Fix automatic loading of styles!
@@ -506,6 +509,7 @@ namespace Starcounter.InstallerWPF {
             System.Windows.Forms.Screen screen = this.GetCurrentScreen();
 
             MainWindow mainWindow = new MainWindow();
+            mainWindow.FinishedMessageInUnattendedMode = this.finishedMessage;
             mainWindow.DefaultSetupOptions = this.setupOptions;
             mainWindow.Configuration.Unattended = this.unattended;
             mainWindow.Configuration.IsUpgrade = this.isUpgrade;
@@ -553,7 +557,6 @@ namespace Starcounter.InstallerWPF {
         }
 
         private void OnError(Exception e) {
-
 
             // Send the tracking error before we close down.
             Dispatcher disp = Dispatcher.FromThread(Thread.CurrentThread);
@@ -668,7 +671,8 @@ namespace Starcounter.InstallerWPF {
         // unattended setup is not the same as silent, silent should not show any qui.
         Boolean unattended = false;
         SetupOptions setupOptions = SetupOptions.None;
-
+        // Message boxtext after upgrade
+        string finishedMessage = string.Empty;
         // Keep settings when uninstalling (set to true when updating starcounter)
         Boolean isUpgrade = false;
 
@@ -711,6 +715,11 @@ namespace Starcounter.InstallerWPF {
                     args = args.Where(w => w != args[i]).ToArray(); // This argument can not be passed along to RunInternalSetup(...)
                     i--;
                     this.setupOptions = SetupOptions.Uninstall;
+                }
+                else if (param.Equals("install", StringComparison.InvariantCultureIgnoreCase)) {
+                    args = args.Where(w => w != args[i]).ToArray(); // This argument can not be passed along to RunInternalSetup(...)
+                    i--;
+                    this.setupOptions = SetupOptions.Install;
                 }
                 else if (param.Equals("upgrade", StringComparison.InvariantCultureIgnoreCase)) {
                     args = args.Where(w => w != args[i]).ToArray(); // This argument can not be passed along to RunInternalSetup(...)